fix env variable name

+++ zookeeper/run.sh
@@ -3,6 +3,9 @@
 if [[ -z "$ZOOKEEPER_ID" ]]; then
   exit 1
 fi
+if [[ -z "$ZOOKEEPER_DATA_DIR" ]]; then
+  exit 1
+fi
 
 echo "$ZOOKEEPER_ID" > /data/zookeeper/myid
 
@@ -17,7 +20,7 @@ autopurge.snapRetainCount=2
 autopurge.purgeInterval=1
 EOF
 
-if [[ -z "$ZOOKEEPER_ID" ]] ; then
+if [[ -z "$ZOOKEEPER_SERVERS" ]] ; then
   export ZOOKEEPER_SERVERS=localhost:2888:3888
 fi
 printf '%s' "$ZOOKEEPER_SERVERS" | awk 'BEGIN { RS = "," }; { printf "server.%i=%s\n", NR, $0 }' >> /opt/zookeeper/conf/zoo.cfg
